+++ src/card/sets/y2018/mod.rs
@@ -1,6 +1,7 @@
 //! Card sets first released in 2018.
 
 pub(super) mod battlebond;
+pub(super) mod commander_2018;
 pub(super) mod core_set_2019;
 pub(super) mod dominaria;